Fix docstring and silence package import warning
Old docstrings were copy-pasted in some places that causes mismatches
with the current functionalities' descriptions. Now everything inside
the ``discretization`` module is properly documented.

Module ``discretization`` provides a placeholder for the
SpearmanDiscretizer that will be implemented in the near future.
Keeping it the way it was caused raising DevelopmentStageWarning
everytime the package was imported.